Added ShaderLibrary

+++ Havoc/Src/Havoc/Core/Application.h
@@ -37,7 +37,7 @@ namespace Havoc
 
 		LayerStack m_LayerStack;
 
-		float m_LastFrameTime = 0.0f;
+		double m_LastFrameTime = 0.0f;
 
 	private:
 		static Application* s_Instance;
